chore(RHINENG-3078): refactor RulesTable to make the component easy to maintain
