+++ src/PresentationalComponents/RulesTable/helpers.js
@@ -31,8 +31,9 @@ import RuleLabels from '../Labels/RuleLabels';
 import CategoryLabel from '../Labels/CategoryLabel';
 
 import { formatMessages, mapContentToValues } from '../../Utilities/intlHelper';
-import { ruleResolutionRisk } from '../Common/Tables';
 import { conditionalFilterType } from '@redhat-cloud-services/frontend-components/ConditionalFilter';
+import { ruleResolutionRisk, pruneFilters } from '../Common/Tables';
+import { cellWidth, fitContent, sortable } from '@patternfly/react-table';
 
 export const emptyRows = (filters, toggleRulesDisabled) => [
   {
@@ -185,6 +186,7 @@ export const hideReports = async (
     });
   }
 };
+
 export const removeFilterParam = (
   param,
   filters,
@@ -518,3 +520,91 @@ export const buildRows = (
 
   return rows;
 };
+
+export const getColumns = (intl) => [
+  {
+    title: intl.formatMessage(messages.name),
+    transforms: [sortable, cellWidth(40)],
+  },
+  {
+    title: intl.formatMessage(messages.modified),
+    transforms: [sortable, fitContent],
+  },
+  {
+    title: intl.formatMessage(messages.category),
+    transforms: [sortable, fitContent],
+  },
+  {
+    title: intl.formatMessage(messages.totalRisk),
+    transforms: [sortable, fitContent],
+  },
+  {
+    title: intl.formatMessage(messages.systems),
+    transforms: [sortable, fitContent],
+  },
+  {
+    title: intl.formatMessage(messages.remediation),
+    transforms: [sortable, fitContent],
+  },
+];
+
+const buildFilterChips = (filters) => {
+  const localFilters = { ...filters };
+  delete localFilters.topic;
+  delete localFilters.sort;
+  delete localFilters.offset;
+  delete localFilters.limit;
+
+  return pruneFilters(localFilters, FC);
+};
+
+export const sortIndices = {
+  1: 'description',
+  2: 'publish_date',
+  3: 'category',
+  4: 'total_risk',
+  5: 'impacted_count',
+  6: 'playbook_count',
+};
+
+export const getActiveFiltersConfig = (
+  filters,
+  intl,
+  setSearchText,
+  setFilters
+) => ({
+  deleteTitle: intl.formatMessage(messages.resetFilters),
+  filters: buildFilterChips(filters),
+  showDeleteButton: true,
+  onDelete: (_event, itemsToRemove, isAll) => {
+    if (isAll) {
+      setSearchText('');
+      setFilters({
+        ...(filters.topic && { topic: filters.topic }),
+        impacting: ['true'],
+        rule_status: 'enabled',
+        limit: filters.limit,
+        offset: filters.offset,
+        pathway: filters.pathway,
+      });
+    } else {
+      itemsToRemove.map((item) => {
+        const newFilter = {
+          [item.urlParam]: Array.isArray(filters[item.urlParam])
+            ? filters[item.urlParam].filter(
+                (value) => String(value) !== String(item.chips[0].value)
+              )
+            : '',
+        };
+        newFilter[item.urlParam].length > 0
+          ? setFilters({ ...filters, ...newFilter })
+          : removeFilterParam(
+              item.urlParam,
+              filters,
+              setFilters,
+              setSearchText
+            );
+      });
+    }
+  },
+});
